heredoc free syuusei

+++ srcs/my_redirect2.c
@@ -6,7 +6,7 @@
 /*   By: wyohei <wyohei@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/14 23:45:23 by wyohei            #+#    #+#             */
-/*   Updated: 2022/02/23 20:05:41 by wyohei           ###   ########.fr       */
+/*   Updated: 2022/02/23 20:55:41 by wyohei           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,9 @@ void	hedoc_fd(t_redirect *redirect)
 		redirect->fd = open(redirect->tmpfile, O_RDONLY);
 		my_dup2(redirect->fd, 0);
 		my_close(redirect->fd);
+	}
+	if (redirect->tmpfile)
+	{
 		unlink(redirect->tmpfile);
 		free(redirect->tmpfile);
 	}
